clean up headers
